mv unit test to integration (need contracts deployment)
